refactor(axdevice,axvm): move LoongArch ACPI table generation from fw_cfg into axvm boot layer
